pyproject.toml
[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "devpi-poetry"
description = "devpi-poetry: devpi plugin to work with poetry"
version = "0.0.1"
authors = ['Andreas Sodeur <andreas.sodeur@rwe.com>']
license = "0BSD"

packages = [
    {include = "devpi_poetry"}
]
include = ["poetry.lock"]

[tool.poetry.dependencies]
python = ">=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*"
devpi-client = ">=4.3.0,<6.0.0"
pluggy = ">=0.6.0,<1.0"
poetry = "poetry==1.0.0a4+rwe"
tomlkit = ">=0.5.3,<1.0.0"

[tool.poetry.plugins.devpi_client]
devpi-poetry =  "devpi_poetry"
